Start of unit tests for split_args function, moved split_args to utils since not needed by modules (so far).
